General-to-Specific (GETS) Modelling

For an overview of the gets package, see gets-package. Here, documentation of generic functions for GETS modelling is provided. Note that gets.arx is a convenience wrapper to getsm and getsv. For specific GETS methods for lm, logitx and isat models, see gets.lm, gets.logitx and gets.isat, respectively.

Usage

gets(x, ...)

# S3 method for class 'arx'
gets(x, spec=NULL, ...)

Arguments

x

an object to be subjected to GETS modelling

spec

NULL (default), "mean" or "variance". If "mean", then getsm is called. If "variance", then getsv is called. If NULL, then it is automatically determined whether GETS-modelling of the mean or log-variance specification should be undertaken.

...

further arguments passed to or from other methods

Details

gets.arx is a convenience wrapper to getsm and getsv.

Author

Genaro Sucarrat, http://www.sucarrat.net/

See also