Recursive estimation

Recursive estimation of coefficients and standard errors

Usage

recursive(object, spec="mean", std.errors=TRUE, from=40, tol=1e-07,
  LAPACK=FALSE, plot=TRUE, return=TRUE)

Arguments

object

an arx, gets or isat object

spec

'mean' or 'variance'. If 'mean' (default), the the recursive estimates of the mean-equation are estimated

std.errors

logical. If TRUE (default), then the coefficient standard errors are also computed

from

integer. The starting point of the recursion

tol

numeric. The tolerance for linear dependency among regressors

LAPACK

logical, TRUE or FALSE (default). If true use LAPACK otherwise use LINPACK, see qr function

plot

NULL or logical. If TRUE, then the recursive coefficient estimates are plotted. If NULL (default), then the value set by options determines whether a plot is produced or not.

return

logical. If TRUE (default), then the recursive estimates are returned in a list

Value

If return=TRUE, then a list is returned with the following components:

estimates

a zoo matrix with the recursive estimates

standard.errors

a zoo matrix with the standard errors

Author

Genaro Sucarrat, http://www.sucarrat.net/

See also

Examples

##generate random variates, estimate model:
y <- rnorm(100)
mX <- matrix(rnorm(4*100), 100, 4)
mymodel <- arx(y, mc=TRUE, mxreg=mX)

##compute recursive estimates and plot them:
recursive(mymodel)
#> $estimates
#>          mconst       mxreg1       mxreg2       mxreg3      mxreg4
#> 1            NA           NA           NA           NA          NA
#> 2            NA           NA           NA           NA          NA
#> 3            NA           NA           NA           NA          NA
#> 4            NA           NA           NA           NA          NA
#> 5            NA           NA           NA           NA          NA
#> 6            NA           NA           NA           NA          NA
#> 7            NA           NA           NA           NA          NA
#> 8            NA           NA           NA           NA          NA
#> 9            NA           NA           NA           NA          NA
#> 10           NA           NA           NA           NA          NA
#> 11           NA           NA           NA           NA          NA
#> 12           NA           NA           NA           NA          NA
#> 13           NA           NA           NA           NA          NA
#> 14           NA           NA           NA           NA          NA
#> 15           NA           NA           NA           NA          NA
#> 16           NA           NA           NA           NA          NA
#> 17           NA           NA           NA           NA          NA
#> 18           NA           NA           NA           NA          NA
#> 19           NA           NA           NA           NA          NA
#> 20           NA           NA           NA           NA          NA
#> 21           NA           NA           NA           NA          NA
#> 22           NA           NA           NA           NA          NA
#> 23           NA           NA           NA           NA          NA
#> 24           NA           NA           NA           NA          NA
#> 25           NA           NA           NA           NA          NA
#> 26           NA           NA           NA           NA          NA
#> 27           NA           NA           NA           NA          NA
#> 28           NA           NA           NA           NA          NA
#> 29           NA           NA           NA           NA          NA
#> 30           NA           NA           NA           NA          NA
#> 31           NA           NA           NA           NA          NA
#> 32           NA           NA           NA           NA          NA
#> 33           NA           NA           NA           NA          NA
#> 34           NA           NA           NA           NA          NA
#> 35           NA           NA           NA           NA          NA
#> 36           NA           NA           NA           NA          NA
#> 37           NA           NA           NA           NA          NA
#> 38           NA           NA           NA           NA          NA
#> 39           NA           NA           NA           NA          NA
#> 40  -0.12420234  0.109511829 -0.188579542 -0.120123006  0.01037086
#> 41  -0.16503922  0.068719036 -0.103940114 -0.115599829 -0.05675442
#> 42  -0.19259684  0.094384503 -0.097323181 -0.101808722 -0.01699945
#> 43  -0.19742066  0.099864933 -0.094790079 -0.094595478 -0.01661601
#> 44  -0.19771561  0.100080609 -0.094662610 -0.095074790 -0.01681273
#> 45  -0.18207229  0.093991806 -0.099390176 -0.087043382 -0.02364039
#> 46  -0.17728524  0.092666274 -0.101510258 -0.094599184 -0.02760881
#> 47  -0.20998582  0.108572541 -0.147767294 -0.095795511 -0.01676736
#> 48  -0.20558044  0.101943902 -0.147065697 -0.100334761 -0.02215299
#> 49  -0.19368943  0.097281913 -0.146688039 -0.105228205 -0.03970630
#> 50  -0.21144013  0.083821654 -0.115495133 -0.090882903 -0.06358777
#> 51  -0.18848756  0.059270146 -0.102568244 -0.059303955 -0.06344370
#> 52  -0.16159184  0.048215408 -0.059632813 -0.056236446 -0.04011384
#> 53  -0.17759046  0.064797173 -0.051229282 -0.031114322 -0.06909971
#> 54  -0.18307527  0.050694084 -0.050188596 -0.025894225 -0.07492311
#> 55  -0.20455807  0.053140752 -0.060346619 -0.027213193 -0.07778139
#> 56  -0.19102033  0.060190829 -0.056027459 -0.019202645 -0.08380436
#> 57  -0.17827491  0.026042989 -0.062744381 -0.027576603 -0.09484493
#> 58  -0.16551859  0.039371161 -0.073909561 -0.007488921 -0.09005842
#> 59  -0.17598667  0.042865486 -0.079022603 -0.011738789 -0.08820317
#> 60  -0.16685383  0.040052468 -0.058589877 -0.004391447 -0.09118327
#> 61  -0.16105645  0.024736154 -0.068466241 -0.011148730 -0.09873518
#> 62  -0.16426859  0.022407101 -0.056688702 -0.005812840 -0.08801453
#> 63  -0.17620817 -0.007219544 -0.072324202 -0.014391950 -0.08494350
#> 64  -0.15894121 -0.000686058 -0.073147525 -0.013896363 -0.10373407
#> 65  -0.16685625 -0.001552078 -0.031522933 -0.018862890 -0.09843476
#> 66  -0.18007083 -0.023350743 -0.034660992 -0.012720233 -0.09183593
#> 67  -0.14974868 -0.027870631  0.024332673 -0.005526346 -0.09360111
#> 68  -0.14431793 -0.033122108  0.023810013 -0.008054852 -0.09227827
#> 69  -0.14052726 -0.037940960  0.025459060 -0.009673020 -0.09203155
#> 70  -0.14520900 -0.032925869  0.024126167 -0.008327806 -0.09546471
#> 71  -0.13774311 -0.034476648  0.023481941  0.002737406 -0.09591747
#> 72  -0.13507685 -0.030521527  0.021982527  0.006599422 -0.09947896
#> 73  -0.14335951 -0.041976303  0.027096624 -0.010265026 -0.09171465
#> 74  -0.12637862 -0.056714162  0.035849963 -0.035193813 -0.08462428
#> 75  -0.12809539 -0.055138363  0.035570291 -0.034998250 -0.08429681
#> 76  -0.12600944 -0.052993320  0.032430693 -0.036768766 -0.08895978
#> 77  -0.14291157 -0.041823426 -0.003493853 -0.037357305 -0.08093703
#> 78  -0.14619974 -0.040999386 -0.004457617 -0.043380214 -0.08236607
#> 79  -0.12931015 -0.044385807 -0.010651924 -0.028254376 -0.07736986
#> 80  -0.13422374 -0.045402969 -0.013557451 -0.034035940 -0.07486581
#> 81  -0.12762825 -0.037279519 -0.015064745 -0.034565663 -0.06908263
#> 82  -0.12838074 -0.037352266 -0.016182357 -0.035173541 -0.06956643
#> 83  -0.13941687 -0.039697634 -0.036637095 -0.045542495 -0.07701198
#> 84  -0.13216443 -0.042135584 -0.026053280 -0.043072000 -0.07387030
#> 85  -0.13535265 -0.042605219 -0.024347275 -0.044388490 -0.07558370
#> 86  -0.14361794 -0.025155366 -0.028687789 -0.037718989 -0.06787485
#> 87  -0.11405560 -0.022759621 -0.065924164  0.018161232 -0.05605461
#> 88  -0.11479048 -0.023847199 -0.066088545  0.018933008 -0.05484216
#> 89  -0.12477564 -0.020429768 -0.054473589  0.016337472 -0.04978723
#> 90  -0.12543128 -0.020352047 -0.052338234  0.015360133 -0.05037378
#> 91  -0.12001427 -0.020127239 -0.056487476  0.012686885 -0.04806313
#> 92  -0.09654717 -0.019287312 -0.037602366 -0.025949096 -0.02565188
#> 93  -0.09817875 -0.018671449 -0.035722553 -0.028352541 -0.01963242
#> 94  -0.10924634 -0.014469354 -0.025779103 -0.018943200 -0.01680497
#> 95  -0.10004666 -0.003474335 -0.017138314 -0.023025355 -0.02387179
#> 96  -0.08678146  0.031072171 -0.040475077 -0.027433287 -0.01885351
#> 97  -0.09447758  0.038828578 -0.035266328 -0.027543409 -0.02261122
#> 98  -0.08835986  0.043837388 -0.052609578 -0.027947908 -0.01824970
#> 99  -0.10067859  0.050416573 -0.051770479 -0.019398250 -0.03319157
#> 100 -0.11244332  0.062830328 -0.058597935 -0.012683886 -0.05142500
#> 
#> $standard.errors
#>       mconstSE   mxreg1SE  mxreg2SE   mxreg3SE   mxreg4SE
#> 1           NA         NA        NA         NA         NA
#> 2           NA         NA        NA         NA         NA
#> 3           NA         NA        NA         NA         NA
#> 4           NA         NA        NA         NA         NA
#> 5           NA         NA        NA         NA         NA
#> 6           NA         NA        NA         NA         NA
#> 7           NA         NA        NA         NA         NA
#> 8           NA         NA        NA         NA         NA
#> 9           NA         NA        NA         NA         NA
#> 10          NA         NA        NA         NA         NA
#> 11          NA         NA        NA         NA         NA
#> 12          NA         NA        NA         NA         NA
#> 13          NA         NA        NA         NA         NA
#> 14          NA         NA        NA         NA         NA
#> 15          NA         NA        NA         NA         NA
#> 16          NA         NA        NA         NA         NA
#> 17          NA         NA        NA         NA         NA
#> 18          NA         NA        NA         NA         NA
#> 19          NA         NA        NA         NA         NA
#> 20          NA         NA        NA         NA         NA
#> 21          NA         NA        NA         NA         NA
#> 22          NA         NA        NA         NA         NA
#> 23          NA         NA        NA         NA         NA
#> 24          NA         NA        NA         NA         NA
#> 25          NA         NA        NA         NA         NA
#> 26          NA         NA        NA         NA         NA
#> 27          NA         NA        NA         NA         NA
#> 28          NA         NA        NA         NA         NA
#> 29          NA         NA        NA         NA         NA
#> 30          NA         NA        NA         NA         NA
#> 31          NA         NA        NA         NA         NA
#> 32          NA         NA        NA         NA         NA
#> 33          NA         NA        NA         NA         NA
#> 34          NA         NA        NA         NA         NA
#> 35          NA         NA        NA         NA         NA
#> 36          NA         NA        NA         NA         NA
#> 37          NA         NA        NA         NA         NA
#> 38          NA         NA        NA         NA         NA
#> 39          NA         NA        NA         NA         NA
#> 40  0.14857005 0.14801361 0.1732117 0.14533095 0.14072922
#> 41  0.16046009 0.15985472 0.1849846 0.15775625 0.15041211
#> 42  0.15739460 0.15715063 0.1845012 0.15676111 0.14391586
#> 43  0.15350280 0.15270457 0.1817109 0.15049272 0.14207122
#> 44  0.14962394 0.14971705 0.1790686 0.14337977 0.13932868
#> 45  0.14702281 0.14851889 0.1778080 0.14201804 0.13810880
#> 46  0.14392811 0.14669331 0.1755235 0.13675440 0.13549198
#> 47  0.14195145 0.14678847 0.1719717 0.13741413 0.13583953
#> 48  0.13865590 0.14138836 0.1700056 0.13400332 0.13164607
#> 49  0.13651282 0.14026468 0.1688740 0.13289894 0.12795033
#> 50  0.13636432 0.14049144 0.1676648 0.13299987 0.12700165
#> 51  0.13403190 0.13791221 0.1669141 0.12859026 0.12685528
#> 52  0.13234161 0.13799877 0.1630977 0.12896328 0.12557703
#> 53  0.13136184 0.13698940 0.1628736 0.12647775 0.12216748
#> 54  0.12943397 0.13077765 0.1614259 0.12463877 0.12015408
#> 55  0.12861728 0.13125449 0.1618021 0.12510403 0.12058273
#> 56  0.12678281 0.13034712 0.1609974 0.12410327 0.11979118
#> 57  0.12508200 0.12148600 0.1600539 0.12306160 0.11835907
#> 58  0.12372521 0.12001955 0.1589787 0.12022270 0.11784838
#> 59  0.12175054 0.11916226 0.1578018 0.11929672 0.11710488
#> 60  0.11900167 0.11809938 0.1493942 0.11719522 0.11603204
#> 61  0.11748079 0.11241834 0.1467666 0.11542943 0.11402890
#> 62  0.11617545 0.11132724 0.1413974 0.11343624 0.10860948
#> 63  0.11446932 0.10281949 0.1390869 0.11231413 0.10806159
#> 64  0.11443625 0.10334681 0.1399609 0.11302035 0.10779543
#> 65  0.11395074 0.10320012 0.1321569 0.11273408 0.10749103
#> 66  0.11344441 0.10134174 0.1323036 0.11274262 0.10746218
#> 67  0.11370445 0.10282542 0.1295570 0.11435173 0.10906812
#> 68  0.11204126 0.10120549 0.1286681 0.11338238 0.10827103
#> 69  0.11030016 0.09880802 0.1275774 0.11238917 0.10747680
#> 70  0.10862738 0.09695551 0.1266357 0.11154199 0.10623383
#> 71  0.10718780 0.09639697 0.1259548 0.10909881 0.10566401
#> 72  0.10585583 0.09430220 0.1249111 0.10713944 0.10387443
#> 73  0.10502681 0.09289655 0.1243855 0.10465562 0.10310631
#> 74  0.10463925 0.09258202 0.1247490 0.10326190 0.10341480
#> 75  0.10320259 0.09126339 0.1238571 0.10252732 0.10266260
#> 76  0.10219265 0.09026807 0.1224194 0.10161194 0.10032791
#> 77  0.10166616 0.09015869 0.1194073 0.10199448 0.10049806
#> 78  0.10019364 0.08952150 0.1185802 0.09857285 0.09969621
#> 79  0.10040093 0.09024870 0.1195090 0.09888474 0.10048196
#> 80  0.09928144 0.08974111 0.1187021 0.09753459 0.09979614
#> 81  0.09818051 0.08822970 0.1181491 0.09709958 0.09884913
#> 82  0.09674831 0.08764876 0.1159219 0.09594551 0.09788192
#> 83  0.09592902 0.08753427 0.1137055 0.09520774 0.09746561
#> 84  0.09489947 0.08712352 0.1120733 0.09477287 0.09697697
#> 85  0.09381806 0.08661813 0.1113079 0.09414427 0.09627504
#> 86  0.09304910 0.08367081 0.1109338 0.09357297 0.09558684
#> 87  0.09817157 0.08867735 0.1169785 0.09757146 0.10124041
#> 88  0.09712127 0.08702326 0.1162566 0.09647407 0.09941188
#> 89  0.09677268 0.08704989 0.1158689 0.09653758 0.09939762
#> 90  0.09583083 0.08653352 0.1118309 0.09513567 0.09852316
#> 91  0.09490615 0.08617435 0.1111030 0.09461264 0.09802213
#> 92  0.09545028 0.08742446 0.1122560 0.09370157 0.09870950
#> 93  0.09469908 0.08691820 0.1113838 0.09266961 0.09498607
#> 94  0.09448296 0.08706293 0.1113508 0.09256648 0.09519219
#> 95  0.09398926 0.08631124 0.1109714 0.09244903 0.09489335
#> 96  0.09453525 0.08454064 0.1111155 0.09328770 0.09574502
#> 97  0.09381852 0.08377479 0.1106782 0.09308880 0.09541996
#> 98  0.09333769 0.08338895 0.1083736 0.09291349 0.09508909
#> 99  0.09287602 0.08333268 0.1085546 0.09277143 0.09435394
#> 100 0.09215746 0.08243558 0.1083586 0.09254574 0.09262982
#>