Recursive estimation

Recursive estimation of coefficients and standard errors

Usage

recursive(object, spec="mean", std.errors=TRUE, from=40, tol=1e-07,
  LAPACK=FALSE, plot=TRUE, return=TRUE)

Arguments

object

an arx, gets or isat object

spec

'mean' or 'variance'. If 'mean' (default), the the recursive estimates of the mean-equation are estimated

std.errors

logical. If TRUE (default), then the coefficient standard errors are also computed

from

integer. The starting point of the recursion

tol

numeric. The tolerance for linear dependency among regressors

LAPACK

logical, TRUE or FALSE (default). If true use LAPACK otherwise use LINPACK, see qr function

plot

NULL or logical. If TRUE, then the recursive coefficient estimates are plotted. If NULL (default), then the value set by options determines whether a plot is produced or not.

return

logical. If TRUE (default), then the recursive estimates are returned in a list

Value

If return=TRUE, then a list is returned with the following components:

estimates

a zoo matrix with the recursive estimates

standard.errors

a zoo matrix with the standard errors

Author

Genaro Sucarrat, http://www.sucarrat.net/

See also

Examples

##generate random variates, estimate model:
y <- rnorm(100)
mX <- matrix(rnorm(4*100), 100, 4)
mymodel <- arx(y, mc=TRUE, mxreg=mX)

##compute recursive estimates and plot them:
recursive(mymodel)
#> $estimates
#>          mconst       mxreg1       mxreg2        mxreg3       mxreg4
#> 1            NA           NA           NA            NA           NA
#> 2            NA           NA           NA            NA           NA
#> 3            NA           NA           NA            NA           NA
#> 4            NA           NA           NA            NA           NA
#> 5            NA           NA           NA            NA           NA
#> 6            NA           NA           NA            NA           NA
#> 7            NA           NA           NA            NA           NA
#> 8            NA           NA           NA            NA           NA
#> 9            NA           NA           NA            NA           NA
#> 10           NA           NA           NA            NA           NA
#> 11           NA           NA           NA            NA           NA
#> 12           NA           NA           NA            NA           NA
#> 13           NA           NA           NA            NA           NA
#> 14           NA           NA           NA            NA           NA
#> 15           NA           NA           NA            NA           NA
#> 16           NA           NA           NA            NA           NA
#> 17           NA           NA           NA            NA           NA
#> 18           NA           NA           NA            NA           NA
#> 19           NA           NA           NA            NA           NA
#> 20           NA           NA           NA            NA           NA
#> 21           NA           NA           NA            NA           NA
#> 22           NA           NA           NA            NA           NA
#> 23           NA           NA           NA            NA           NA
#> 24           NA           NA           NA            NA           NA
#> 25           NA           NA           NA            NA           NA
#> 26           NA           NA           NA            NA           NA
#> 27           NA           NA           NA            NA           NA
#> 28           NA           NA           NA            NA           NA
#> 29           NA           NA           NA            NA           NA
#> 30           NA           NA           NA            NA           NA
#> 31           NA           NA           NA            NA           NA
#> 32           NA           NA           NA            NA           NA
#> 33           NA           NA           NA            NA           NA
#> 34           NA           NA           NA            NA           NA
#> 35           NA           NA           NA            NA           NA
#> 36           NA           NA           NA            NA           NA
#> 37           NA           NA           NA            NA           NA
#> 38           NA           NA           NA            NA           NA
#> 39           NA           NA           NA            NA           NA
#> 40  -0.17410857  0.190448765 -0.153449653 -0.0522895601  0.058887732
#> 41  -0.14882690  0.153260673 -0.146047535 -0.1047398575  0.032680362
#> 42  -0.15072717  0.145478545 -0.137085832 -0.1134450862  0.032723432
#> 43  -0.18840575  0.101539981 -0.059390964 -0.1094608741 -0.035863855
#> 44  -0.21024965  0.120348633 -0.058722713 -0.0980135338 -0.002256577
#> 45  -0.21323119  0.123617196 -0.057646429 -0.0930626244 -0.002318326
#> 46  -0.21252232  0.123137037 -0.057861580 -0.0918229426 -0.001767207
#> 47  -0.19709350  0.118402643 -0.060522498 -0.0829735313 -0.007912064
#> 48  -0.19190723  0.117551789 -0.062159401 -0.0917128657 -0.012161324
#> 49  -0.22492257  0.133719140 -0.107955999 -0.0931048928 -0.001081988
#> 50  -0.21970475  0.125922987 -0.105851956 -0.0989094625 -0.007635897
#> 51  -0.20789484  0.122296777 -0.103668869 -0.1038978059 -0.025533388
#> 52  -0.22443352  0.107825340 -0.075584016 -0.0899837833 -0.049579685
#> 53  -0.20147540  0.083883443 -0.059924651 -0.0570270442 -0.048563049
#> 54  -0.17471262  0.071894532 -0.020605785 -0.0541021314 -0.026682712
#> 55  -0.18939286  0.086526389 -0.014471721 -0.0299761119 -0.055588334
#> 56  -0.19568222  0.071852188 -0.012451541 -0.0242442868 -0.061532958
#> 57  -0.21680395  0.073889363 -0.023024952 -0.0256589009 -0.064606397
#> 58  -0.20327991  0.081047587 -0.018576508 -0.0176060503 -0.070595954
#> 59  -0.19081311  0.045551803 -0.021201017 -0.0265956603 -0.081132708
#> 60  -0.17793803  0.059383424 -0.031818824 -0.0061725211 -0.076058810
#> 61  -0.18795547  0.062499637 -0.037291931 -0.0103749163 -0.074441488
#> 62  -0.17949373  0.059554889 -0.019865482 -0.0037695980 -0.077420406
#> 63  -0.17347333  0.042243439 -0.028982243 -0.0120160223 -0.085609797
#> 64  -0.17552068  0.040011935 -0.021196525 -0.0080316479 -0.077952706
#> 65  -0.18984295  0.008365631 -0.035966722 -0.0177083213 -0.073693065
#> 66  -0.17278340  0.015548291 -0.035755100 -0.0172387477 -0.092415820
#> 67  -0.17861141  0.013320344 -0.001932573 -0.0213749248 -0.088697403
#> 68  -0.19229903 -0.008480920 -0.004456449 -0.0151903709 -0.081775713
#> 69  -0.16180085 -0.013917554  0.050931493 -0.0078541600 -0.084219726
#> 70  -0.15615127 -0.019328808  0.050905285 -0.0106361027 -0.082555027
#> 71  -0.15208362 -0.024523080  0.053079173 -0.0124523842 -0.082140145
#> 72  -0.15626690 -0.020051689  0.051579047 -0.0112028600 -0.085421518
#> 73  -0.14866582 -0.021443599  0.051400786  0.0002967331 -0.085715923
#> 74  -0.14590836 -0.017274555  0.049934088  0.0043160184 -0.089407390
#> 75  -0.15401621 -0.028672117  0.054680836 -0.0122682326 -0.081822352
#> 76  -0.13722884 -0.043114115  0.064284494 -0.0375020507 -0.074265266
#> 77  -0.13851610 -0.041947671  0.063953891 -0.0373459976 -0.074056379
#> 78  -0.13627371 -0.039342466  0.060711338 -0.0393812287 -0.079237114
#> 79  -0.15365307 -0.027687891  0.025092150 -0.0401556303 -0.070856257
#> 80  -0.15676569 -0.026939251  0.024081269 -0.0458948455 -0.072249109
#> 81  -0.14002488 -0.029788959  0.018991621 -0.0306090568 -0.066734829
#> 82  -0.14490004 -0.030818323  0.016070574 -0.0363603737 -0.064255983
#> 83  -0.13834337 -0.022932591  0.014428323 -0.0368809329 -0.058641584
#> 84  -0.13949938 -0.023002648  0.012835130 -0.0377979633 -0.059337853
#> 85  -0.15148624 -0.024884695 -0.007407831 -0.0487476624 -0.066718947
#> 86  -0.14432511 -0.027386321  0.002599429 -0.0462941343 -0.063751095
#> 87  -0.14723613 -0.027907362  0.004030595 -0.0475047422 -0.065408409
#> 88  -0.15448750 -0.012129356 -0.001025344 -0.0412056274 -0.058582967
#> 89  -0.12573796 -0.008300473 -0.035473787  0.0148233917 -0.045617601
#> 90  -0.12650812 -0.009438336 -0.035637056  0.0156286979 -0.044346300
#> 91  -0.13554057 -0.006815963 -0.025816484  0.0132596747 -0.039972844
#> 92  -0.13561766 -0.006814614 -0.025562157  0.0131369372 -0.040054051
#> 93  -0.13009130 -0.006365831 -0.029527868  0.0103064373 -0.037467106
#> 94  -0.10664330 -0.005779726 -0.011476718 -0.0279110798 -0.015426933
#> 95  -0.10784895 -0.005420748 -0.010215150 -0.0298253047 -0.010700855
#> 96  -0.11810465 -0.001986182 -0.001695848 -0.0206696383 -0.008384870
#> 97  -0.10891210  0.008582222  0.006314178 -0.0246325403 -0.015481061
#> 98  -0.09561070  0.043205896 -0.016904644 -0.0291113650 -0.010436748
#> 99  -0.10263228  0.050177131 -0.012707668 -0.0291833397 -0.014314483
#> 100 -0.09653695  0.055856204 -0.029979312 -0.0296788806 -0.009428743
#> 
#> $standard.errors
#>       mconstSE   mxreg1SE  mxreg2SE   mxreg3SE   mxreg4SE
#> 1           NA         NA        NA         NA         NA
#> 2           NA         NA        NA         NA         NA
#> 3           NA         NA        NA         NA         NA
#> 4           NA         NA        NA         NA         NA
#> 5           NA         NA        NA         NA         NA
#> 6           NA         NA        NA         NA         NA
#> 7           NA         NA        NA         NA         NA
#> 8           NA         NA        NA         NA         NA
#> 9           NA         NA        NA         NA         NA
#> 10          NA         NA        NA         NA         NA
#> 11          NA         NA        NA         NA         NA
#> 12          NA         NA        NA         NA         NA
#> 13          NA         NA        NA         NA         NA
#> 14          NA         NA        NA         NA         NA
#> 15          NA         NA        NA         NA         NA
#> 16          NA         NA        NA         NA         NA
#> 17          NA         NA        NA         NA         NA
#> 18          NA         NA        NA         NA         NA
#> 19          NA         NA        NA         NA         NA
#> 20          NA         NA        NA         NA         NA
#> 21          NA         NA        NA         NA         NA
#> 22          NA         NA        NA         NA         NA
#> 23          NA         NA        NA         NA         NA
#> 24          NA         NA        NA         NA         NA
#> 25          NA         NA        NA         NA         NA
#> 26          NA         NA        NA         NA         NA
#> 27          NA         NA        NA         NA         NA
#> 28          NA         NA        NA         NA         NA
#> 29          NA         NA        NA         NA         NA
#> 30          NA         NA        NA         NA         NA
#> 31          NA         NA        NA         NA         NA
#> 32          NA         NA        NA         NA         NA
#> 33          NA         NA        NA         NA         NA
#> 34          NA         NA        NA         NA         NA
#> 35          NA         NA        NA         NA         NA
#> 36          NA         NA        NA         NA         NA
#> 37          NA         NA        NA         NA         NA
#> 38          NA         NA        NA         NA         NA
#> 39          NA         NA        NA         NA         NA
#> 40  0.14744569 0.15430534 0.1737861 0.16304430 0.14045044
#> 41  0.14535437 0.14990310 0.1736852 0.15461784 0.13808000
#> 42  0.14293522 0.13971281 0.1619973 0.14256476 0.13624829
#> 43  0.15387124 0.15010232 0.1724852 0.15418497 0.14483405
#> 44  0.15129334 0.14803819 0.1719254 0.15311887 0.13907305
#> 45  0.14797086 0.14442090 0.1696399 0.14722658 0.13735817
#> 46  0.14449920 0.14187356 0.1674280 0.14026331 0.13459677
#> 47  0.14233065 0.14101571 0.1665405 0.13906606 0.13366958
#> 48  0.13962794 0.13946745 0.1646501 0.13413569 0.13139512
#> 49  0.13773258 0.13960131 0.1611927 0.13487052 0.13180341
#> 50  0.13493324 0.13511532 0.1593317 0.13173653 0.12816767
#> 51  0.13310737 0.13425495 0.1584044 0.13080096 0.12485144
#> 52  0.13294033 0.13427081 0.1572914 0.13083065 0.12371822
#> 53  0.13091463 0.13208928 0.1564897 0.12656455 0.12370326
#> 54  0.12904534 0.13199409 0.1528907 0.12686406 0.12246453
#> 55  0.12812755 0.13110038 0.1527060 0.12441246 0.11882463
#> 56  0.12621905 0.12534027 0.1513932 0.12266021 0.11702333
#> 57  0.12543797 0.12581413 0.1517156 0.12312969 0.11744793
#> 58  0.12367451 0.12496063 0.1509978 0.12218366 0.11671576
#> 59  0.12248865 0.11748709 0.1505796 0.12141404 0.11575943
#> 60  0.12122407 0.11604183 0.1496665 0.11868020 0.11530431
#> 61  0.11932936 0.11524513 0.1485071 0.11777504 0.11459903
#> 62  0.11648311 0.11414251 0.1406311 0.11569284 0.11349413
#> 63  0.11533131 0.10941413 0.1389114 0.11412257 0.11193179
#> 64  0.11412898 0.10818728 0.1343619 0.11212434 0.10688109
#> 65  0.11238318 0.10041701 0.1326932 0.11113865 0.10642967
#> 66  0.11243490 0.10095941 0.1335964 0.11189467 0.10624594
#> 67  0.11190913 0.10065302 0.1266975 0.11148170 0.10586504
#> 68  0.11142208 0.09888279 0.1269029 0.11154017 0.10586866
#> 69  0.11145659 0.10020970 0.1242213 0.11301060 0.10733705
#> 70  0.10994482 0.09875251 0.1234329 0.11210151 0.10658352
#> 71  0.10833659 0.09652536 0.1223630 0.11115982 0.10583601
#> 72  0.10674612 0.09477155 0.1214359 0.11033167 0.10457999
#> 73  0.10540199 0.09427322 0.1208364 0.10796897 0.10406279
#> 74  0.10412148 0.09218526 0.1198752 0.10606299 0.10233007
#> 75  0.10331081 0.09076434 0.1193851 0.10361812 0.10158134
#> 76  0.10300161 0.09052217 0.1197372 0.10228346 0.10190928
#> 77  0.10164341 0.08929241 0.1188764 0.10156993 0.10119067
#> 78  0.10071151 0.08828091 0.1176054 0.10068989 0.09897649
#> 79  0.10019961 0.08817245 0.1147068 0.10110946 0.09917065
#> 80  0.09876502 0.08756543 0.1139118 0.09773643 0.09838491
#> 81  0.09904360 0.08833026 0.1148840 0.09810153 0.09920058
#> 82  0.09794850 0.08784428 0.1141098 0.09677419 0.09853732
#> 83  0.09684419 0.08636504 0.1135772 0.09634911 0.09761489
#> 84  0.09539641 0.08581777 0.1115026 0.09521323 0.09669948
#> 85  0.09457454 0.08577070 0.1095243 0.09452199 0.09637477
#> 86  0.09352756 0.08535549 0.1079446 0.09408567 0.09589625
#> 87  0.09248291 0.08486043 0.1072326 0.09347057 0.09519584
#> 88  0.09175084 0.08205714 0.1067503 0.09286158 0.09452590
#> 89  0.09689630 0.08702328 0.1127075 0.09691852 0.10017406
#> 90  0.09586614 0.08538504 0.1120292 0.09584508 0.09836397
#> 91  0.09551353 0.08538617 0.1116794 0.09586113 0.09831805
#> 92  0.09464456 0.08489397 0.1080353 0.09450215 0.09740804
#> 93  0.09378068 0.08456308 0.1073941 0.09400703 0.09692654
#> 94  0.09425873 0.08575621 0.1084735 0.09307583 0.09757233
#> 95  0.09354985 0.08527670 0.1076983 0.09205621 0.09398163
#> 96  0.09332852 0.08539976 0.1076711 0.09191025 0.09415212
#> 97  0.09280728 0.08466401 0.1073113 0.09178465 0.09383012
#> 98  0.09334509 0.08286105 0.1074183 0.09261969 0.09467510
#> 99  0.09265411 0.08214219 0.1070200 0.09240240 0.09431135
#> 100 0.09224212 0.08175124 0.1049233 0.09226604 0.09399924
#>