Recursive estimation

Recursive estimation of coefficients and standard errors

Usage

recursive(object, spec="mean", std.errors=TRUE, from=40, tol=1e-07,
  LAPACK=FALSE, plot=TRUE, return=TRUE)

Arguments

object

an arx, gets or isat object

spec

'mean' or 'variance'. If 'mean' (default), the the recursive estimates of the mean-equation are estimated

std.errors

logical. If TRUE (default), then the coefficient standard errors are also computed

from

integer. The starting point of the recursion

tol

numeric. The tolerance for linear dependency among regressors

LAPACK

logical, TRUE or FALSE (default). If true use LAPACK otherwise use LINPACK, see qr function

plot

NULL or logical. If TRUE, then the recursive coefficient estimates are plotted. If NULL (default), then the value set by options determines whether a plot is produced or not.

return

logical. If TRUE (default), then the recursive estimates are returned in a list

Value

If return=TRUE, then a list is returned with the following components:

estimates

a zoo matrix with the recursive estimates

standard.errors

a zoo matrix with the standard errors

Author

Genaro Sucarrat, http://www.sucarrat.net/

See also

Examples

##generate random variates, estimate model:
y <- rnorm(100)
mX <- matrix(rnorm(4*100), 100, 4)
mymodel <- arx(y, mc=TRUE, mxreg=mX)

##compute recursive estimates and plot them:
recursive(mymodel)
#> $estimates
#>         mconst      mxreg1        mxreg2        mxreg3       mxreg4
#> 1           NA          NA            NA            NA           NA
#> 2           NA          NA            NA            NA           NA
#> 3           NA          NA            NA            NA           NA
#> 4           NA          NA            NA            NA           NA
#> 5           NA          NA            NA            NA           NA
#> 6           NA          NA            NA            NA           NA
#> 7           NA          NA            NA            NA           NA
#> 8           NA          NA            NA            NA           NA
#> 9           NA          NA            NA            NA           NA
#> 10          NA          NA            NA            NA           NA
#> 11          NA          NA            NA            NA           NA
#> 12          NA          NA            NA            NA           NA
#> 13          NA          NA            NA            NA           NA
#> 14          NA          NA            NA            NA           NA
#> 15          NA          NA            NA            NA           NA
#> 16          NA          NA            NA            NA           NA
#> 17          NA          NA            NA            NA           NA
#> 18          NA          NA            NA            NA           NA
#> 19          NA          NA            NA            NA           NA
#> 20          NA          NA            NA            NA           NA
#> 21          NA          NA            NA            NA           NA
#> 22          NA          NA            NA            NA           NA
#> 23          NA          NA            NA            NA           NA
#> 24          NA          NA            NA            NA           NA
#> 25          NA          NA            NA            NA           NA
#> 26          NA          NA            NA            NA           NA
#> 27          NA          NA            NA            NA           NA
#> 28          NA          NA            NA            NA           NA
#> 29          NA          NA            NA            NA           NA
#> 30          NA          NA            NA            NA           NA
#> 31          NA          NA            NA            NA           NA
#> 32          NA          NA            NA            NA           NA
#> 33          NA          NA            NA            NA           NA
#> 34          NA          NA            NA            NA           NA
#> 35          NA          NA            NA            NA           NA
#> 36          NA          NA            NA            NA           NA
#> 37          NA          NA            NA            NA           NA
#> 38          NA          NA            NA            NA           NA
#> 39          NA          NA            NA            NA           NA
#> 40  0.24699500 -0.13189805 -0.1256567624 -0.0141403985 -0.091668297
#> 41  0.23613833 -0.12485242 -0.1214241753 -0.0265661733 -0.095997706
#> 42  0.26413567 -0.12677380 -0.1283569495  0.0203831673 -0.122105542
#> 43  0.22005735 -0.12791618 -0.0810483967 -0.0253145453 -0.114561970
#> 44  0.18368002 -0.10060222 -0.0298869306 -0.0355987996 -0.083055899
#> 45  0.16980768 -0.08863411 -0.0635139280 -0.0271968697 -0.067362277
#> 46  0.15858224 -0.08765230 -0.0949998140  0.0184103765 -0.067102558
#> 47  0.13163610 -0.09960908 -0.0549986594 -0.0044727604 -0.055854170
#> 48  0.13312728 -0.10147787 -0.0558130552 -0.0043600995 -0.054690788
#> 49  0.07344893 -0.08862717  0.0109627348  0.0050860725 -0.082230833
#> 50  0.07712083 -0.08661176  0.0036569252  0.0024403131 -0.083813335
#> 51  0.09985008 -0.09617929  0.0022379936  0.0001916866 -0.074470053
#> 52  0.14248037 -0.08695786 -0.0072959030 -0.0210186104 -0.040320220
#> 53  0.17284889 -0.07555503 -0.0183900195 -0.0186595494 -0.011027183
#> 54  0.17969746 -0.07237178 -0.0126965193 -0.0327292822 -0.014397946
#> 55  0.13455353 -0.09913568 -0.0246786014 -0.0980992915 -0.021410065
#> 56  0.12070908 -0.09845154 -0.0247645244 -0.1005030815 -0.022501737
#> 57  0.10901830 -0.09579879 -0.0127156921 -0.1133735560 -0.035753136
#> 58  0.10969302 -0.09879617 -0.0135142084 -0.1160546191 -0.034467229
#> 59  0.10494760 -0.09898549 -0.0117736638 -0.1159058776 -0.038526638
#> 60  0.08538349 -0.11180572 -0.0235971144 -0.1001051161 -0.030860121
#> 61  0.11245626 -0.09598785 -0.0569318083 -0.1163406946 -0.018921706
#> 62  0.13499402 -0.10331257 -0.0735591612 -0.0851266988 -0.006730561
#> 63  0.13681710 -0.09721516 -0.0674341177 -0.0910873953 -0.011034017
#> 64  0.15357364 -0.09272031 -0.0727877053 -0.0838848424 -0.038799619
#> 65  0.11998795 -0.09859214 -0.0391227404 -0.1152442777 -0.060592408
#> 66  0.12889756 -0.08820805 -0.0436461527 -0.1123500619 -0.071957070
#> 67  0.12216182 -0.08093518 -0.0624156943 -0.1010199611 -0.070577347
#> 68  0.13395545 -0.08436566 -0.0594441595 -0.0899335669 -0.066194433
#> 69  0.13587488 -0.08001036 -0.0568275324 -0.0896679615 -0.064692304
#> 70  0.14082541 -0.07970072 -0.0554073470 -0.0952820828 -0.063464395
#> 71  0.14994982 -0.05644433 -0.0434574078 -0.1180022766 -0.075506628
#> 72  0.15163761 -0.05743936 -0.0449580077 -0.1155942921 -0.074248461
#> 73  0.16376066 -0.05801793 -0.0366739041 -0.1217588498 -0.051124368
#> 74  0.14598354 -0.06609392 -0.0225747898 -0.0931308571 -0.070409207
#> 75  0.12988073 -0.06958934  0.0048194113 -0.1040614125 -0.049873931
#> 76  0.15207711 -0.08315046  0.0009944043 -0.0933398323 -0.023170539
#> 77  0.15252580 -0.08337664  0.0044546721 -0.0951589051 -0.023633861
#> 78  0.13889467 -0.04951627 -0.0318862070 -0.0397931188 -0.029521044
#> 79  0.12300713 -0.06287575 -0.0544698091 -0.0411443398 -0.041140831
#> 80  0.11990564 -0.06553847 -0.0545745355 -0.0442937884 -0.041599189
#> 81  0.12860079 -0.07962998 -0.0615450682 -0.0386364717 -0.042235420
#> 82  0.12689718 -0.08152717 -0.0605052301 -0.0385590646 -0.044113633
#> 83  0.13109972 -0.08619786 -0.0627442685 -0.0410503342 -0.041158336
#> 84  0.14488343 -0.07947945 -0.0743125903 -0.0338438449 -0.027800756
#> 85  0.16552252 -0.06357034 -0.0665949781 -0.0232073004  0.020476519
#> 86  0.16226883 -0.06367565 -0.0618483135 -0.0255742284  0.018433465
#> 87  0.16032738 -0.06210375 -0.0646212006 -0.0248201157  0.018381436
#> 88  0.14013001 -0.06204354 -0.0875224143 -0.0405439709  0.061573546
#> 89  0.13611456 -0.06446657 -0.0807169910 -0.0399133730  0.052780368
#> 90  0.13023651 -0.06967477 -0.0853108853 -0.0383735002  0.055734284
#> 91  0.12312951 -0.04900989 -0.1193134279 -0.0473375655  0.033854908
#> 92  0.11751519 -0.04302978 -0.1144956110 -0.0527549944  0.032517867
#> 93  0.12292767 -0.03055433 -0.1055657034 -0.0679071959  0.043908027
#> 94  0.09339570 -0.05444992 -0.0763077439 -0.0791607669  0.042213098
#> 95  0.08953962 -0.05651592 -0.0808404563 -0.0823201365  0.040505068
#> 96  0.08922481 -0.05675273 -0.0809470238 -0.0818028668  0.040491981
#> 97  0.08289654 -0.06313945 -0.0953057757 -0.0859516001  0.042419279
#> 98  0.08157206 -0.06033453 -0.0931876006 -0.0851048472  0.042790754
#> 99  0.08718584 -0.05513504 -0.0960528801 -0.0787809856  0.047697069
#> 100 0.08536875 -0.05468269 -0.0947477579 -0.0798852315  0.047911547
#> 
#> $standard.errors
#>      mconstSE   mxreg1SE  mxreg2SE  mxreg3SE   mxreg4SE
#> 1          NA         NA        NA        NA         NA
#> 2          NA         NA        NA        NA         NA
#> 3          NA         NA        NA        NA         NA
#> 4          NA         NA        NA        NA         NA
#> 5          NA         NA        NA        NA         NA
#> 6          NA         NA        NA        NA         NA
#> 7          NA         NA        NA        NA         NA
#> 8          NA         NA        NA        NA         NA
#> 9          NA         NA        NA        NA         NA
#> 10         NA         NA        NA        NA         NA
#> 11         NA         NA        NA        NA         NA
#> 12         NA         NA        NA        NA         NA
#> 13         NA         NA        NA        NA         NA
#> 14         NA         NA        NA        NA         NA
#> 15         NA         NA        NA        NA         NA
#> 16         NA         NA        NA        NA         NA
#> 17         NA         NA        NA        NA         NA
#> 18         NA         NA        NA        NA         NA
#> 19         NA         NA        NA        NA         NA
#> 20         NA         NA        NA        NA         NA
#> 21         NA         NA        NA        NA         NA
#> 22         NA         NA        NA        NA         NA
#> 23         NA         NA        NA        NA         NA
#> 24         NA         NA        NA        NA         NA
#> 25         NA         NA        NA        NA         NA
#> 26         NA         NA        NA        NA         NA
#> 27         NA         NA        NA        NA         NA
#> 28         NA         NA        NA        NA         NA
#> 29         NA         NA        NA        NA         NA
#> 30         NA         NA        NA        NA         NA
#> 31         NA         NA        NA        NA         NA
#> 32         NA         NA        NA        NA         NA
#> 33         NA         NA        NA        NA         NA
#> 34         NA         NA        NA        NA         NA
#> 35         NA         NA        NA        NA         NA
#> 36         NA         NA        NA        NA         NA
#> 37         NA         NA        NA        NA         NA
#> 38         NA         NA        NA        NA         NA
#> 39         NA         NA        NA        NA         NA
#> 40  0.1820880 0.14992330 0.2003516 0.1889394 0.15798322
#> 41  0.1751911 0.14566953 0.1971314 0.1807477 0.15510975
#> 42  0.1713279 0.14507743 0.1961828 0.1712525 0.15137920
#> 43  0.1670999 0.14549396 0.1920066 0.1666631 0.15166315
#> 44  0.1660265 0.14510839 0.1893573 0.1678379 0.15088877
#> 45  0.1663094 0.14536536 0.1879323 0.1684047 0.15092917
#> 46  0.1665858 0.14586362 0.1865056 0.1640839 0.15144901
#> 47  0.1611141 0.14398594 0.1764987 0.1598338 0.14968459
#> 48  0.1572021 0.13871037 0.1738902 0.1579591 0.14660974
#> 49  0.1608263 0.14407137 0.1778317 0.1641543 0.15182833
#> 50  0.1569713 0.14179251 0.1682720 0.1612952 0.14975601
#> 51  0.1558251 0.14176510 0.1685647 0.1615671 0.14976974
#> 52  0.1584223 0.14556083 0.1731027 0.1655936 0.15275758
#> 53  0.1565528 0.14561697 0.1733084 0.1660458 0.15095336
#> 54  0.1548802 0.14442743 0.1717103 0.1627310 0.14971475
#> 55  0.1546668 0.14569094 0.1742168 0.1599870 0.15197950
#> 56  0.1525494 0.14493037 0.1733114 0.1591174 0.15118136
#> 57  0.1472698 0.14347668 0.1681213 0.1531310 0.14471415
#> 58  0.1455035 0.13429752 0.1660727 0.1458697 0.14195273
#> 59  0.1428120 0.13311428 0.1644468 0.1445858 0.13964659
#> 60  0.1430335 0.13368280 0.1653423 0.1450929 0.14049301
#> 61  0.1429538 0.13435909 0.1650444 0.1458820 0.14144795
#> 62  0.1399739 0.13370455 0.1633803 0.1405765 0.14030170
#> 63  0.1385232 0.12917561 0.1592125 0.1363073 0.13752544
#> 64  0.1376246 0.12916237 0.1592023 0.1361911 0.13490774
#> 65  0.1360958 0.12988609 0.1581248 0.1349389 0.13472591
#> 66  0.1342705 0.12766098 0.1569676 0.1340361 0.13224518
#> 67  0.1333944 0.12673545 0.1540931 0.1325430 0.13169865
#> 68  0.1314446 0.12601207 0.1532832 0.1307429 0.13088512
#> 69  0.1294656 0.11977417 0.1505615 0.1297139 0.12928198
#> 70  0.1280532 0.11899851 0.1495489 0.1281360 0.12841193
#> 71  0.1279730 0.11726150 0.1493732 0.1266327 0.12813140
#> 72  0.1244214 0.11540772 0.1465079 0.1202842 0.12574092
#> 73  0.1234675 0.11522165 0.1459751 0.1198906 0.12279934
#> 74  0.1238517 0.11601552 0.1468271 0.1192182 0.12306591
#> 75  0.1223505 0.11578014 0.1433707 0.1184139 0.12071755
#> 76  0.1219414 0.11601004 0.1441709 0.1188299 0.11977430
#> 77  0.1209965 0.11518090 0.1379521 0.1162767 0.11883567
#> 78  0.1244335 0.11763754 0.1411205 0.1171749 0.12232755
#> 79  0.1244615 0.11778050 0.1407696 0.1177512 0.12261387
#> 80  0.1224369 0.11606944 0.1398567 0.1156595 0.12179287
#> 81  0.1208052 0.11256647 0.1386083 0.1146452 0.12121802
#> 82  0.1190760 0.11057306 0.1374085 0.1139057 0.11928404
#> 83  0.1178288 0.10918134 0.1364956 0.1130544 0.11832938
#> 84  0.1164717 0.10868899 0.1355526 0.1125225 0.11704567
#> 85  0.1167946 0.10921239 0.1367087 0.1133515 0.11399276
#> 86  0.1148045 0.10855848 0.1335347 0.1119713 0.11279134
#> 87  0.1137454 0.10765471 0.1320738 0.1112558 0.11213045
#> 88  0.1142819 0.10878680 0.1327324 0.1120174 0.11021596
#> 89  0.1128019 0.10786505 0.1298841 0.1113825 0.10529648
#> 90  0.1118586 0.10702905 0.1291007 0.1108983 0.10473905
#> 91  0.1116565 0.10508642 0.1246867 0.1105628 0.10250732
#> 92  0.1101919 0.10346354 0.1234729 0.1091587 0.10195069
#> 93  0.1098230 0.10222941 0.1228141 0.1074892 0.10088353
#> 94  0.1108703 0.10342142 0.1242149 0.1093046 0.10271596
#> 95  0.1098297 0.10276659 0.1229919 0.1084329 0.10211295
#> 96  0.1089196 0.10201634 0.1222840 0.1069979 0.10155060
#> 97  0.1082580 0.10134133 0.1202644 0.1065476 0.10123877
#> 98  0.1070423 0.09769970 0.1181473 0.1057154 0.10064627
#> 99  0.1057126 0.09646264 0.1174240 0.1041460 0.09950597
#> 100 0.1043840 0.09591086 0.1164544 0.1033150 0.09897984
#>